Workshop Slides Written Deliverables
Committing slides used for workshop on alexa skill. 
Note: workshop is only focused on alexa skill and does not include information on libraries. The length of the video is extensive mostly due to it being my first run through and I had moments of rambling but it provides a good first draft. I felt that constructing an alexa skill provided enough detail for a workshop and the extra details on libraries included in the proposal would have made it too long. Attached is the loom link for the video. https://www.loom.com/share/3a962779d2dc4c6388946176d26f51dc